cms栏目专题

+++ jbone-cms/jbone-cms-core/src/main/java/cn/jbone/cms/core/dao/entity/CategoryToc.java
@@ -23,7 +23,7 @@ public class CategoryToc extends BaseEntity {
      * 父级栏目
      */
     @Column(name = "pid")
-    private int pid;
+    private long pid;
 
     /**
      * 栏目标题
